ref(dynamic-sampling): Move functions to the relay-sampling crate
