+++ relay-sampling/src/lib.rs
@@ -878,11 +878,32 @@ impl Display for MatchedRuleIds {
     }
 }
 
+/// Checks whether unsupported rules result in a direct keep of the event or depending on the
+/// type of Relay an ignore of unsupported rules.
+fn check_unsupported_rules(
+    processing_enabled: bool,
+    sampling_config: &SamplingConfig,
+    root_sampling_config: Option<&SamplingConfig>,
+) -> Option<()> {
+    // When we have unsupported rules disable sampling for non processing relays.
+    if sampling_config.has_unsupported_rules()
+        || root_sampling_config.map_or(false, |config| config.has_unsupported_rules())
+    {
+        if !processing_enabled {
+            return None;
+        } else {
+            relay_log::error!("found unsupported rules even as processing relay");
+        }
+    }
+
+    Some(())
+}
+
 /// Returns an iterator of references that chains together and merges rules.
 ///
 /// The chaining logic will take all the non-trace rules from the project and all the trace/unsupported
 /// rules from the root project and concatenate them.
-pub fn merge_rules_from_configs<'a>(
+fn merge_rules_from_configs<'a>(
     sampling_config: &'a SamplingConfig,
     root_sampling_config: Option<&'a SamplingConfig>,
 ) -> impl Iterator<Item = &'a SamplingRule> {
@@ -897,6 +918,50 @@ pub fn merge_rules_from_configs<'a>(
         .filter(|&rule| rule.ty == RuleType::Trace);
 
     event_rules.chain(parent_rules)
+}
+
+/// Gets the sampling match result by creating the merged configuration and matching it against
+/// the sampling configuration.
+pub fn merge_configs_and_match(
+    processing_enabled: bool,
+    sampling_config: &SamplingConfig,
+    root_sampling_config: Option<&SamplingConfig>,
+    dsc: Option<&DynamicSamplingContext>,
+    event: &Event,
+    ip_addr: Option<IpAddr>,
+    now: DateTime<Utc>,
+) -> Option<SamplingMatch> {
+    // We check if there are unsupported rules in any of the two configurations.
+    check_unsupported_rules(processing_enabled, sampling_config, root_sampling_config)?;
+
+    // We perform the rule matching with the multi-matching logic on the merged rules.
+    let rules = merge_rules_from_configs(sampling_config, root_sampling_config);
+    let mut match_result = SamplingMatch::match_against_rules(rules, event, dsc, ip_addr, now)?;
+
+    // If we have a match, we will try to derive the sample rate based on the sampling mode.
+    //
+    // Keep in mind that the sample rate received here has already been derived by the matching
+    // logic, based on multiple matches and decaying functions.
+    //
+    // We also decide to use the sampling mode of the project to which the event belongs.
+    match_result.set_sample_rate(match sampling_config.mode {
+        SamplingMode::Received => match_result.sample_rate,
+        SamplingMode::Total => match dsc {
+            Some(dsc) => dsc.adjusted_sample_rate(match_result.sample_rate),
+            None => match_result.sample_rate,
+        },
+        SamplingMode::Unsupported => {
+            if processing_enabled {
+                relay_log::error!("found unsupported sampling mode even as processing Relay");
+            }
+
+            return None;
+        }
+    });
+
+    // Only if we arrive at this stage, it means that we have found a match and we want to prepare
+    // the data for making the sampling decision.
+    Some(match_result)
 }
 
 /// Represents the specification for sampling an incoming event.
